server/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
celery==5.3.4
django-celery-beat==2.5.0
django-redis==5.4.0
Pillow==10.1.0
python-dotenv==1.0.0
